backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pillow
PyMuPDF
sendgrid
python-dotenv
langchain
langgraph
openai
tiktoken
sendgrid 
reportlab
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib 
google-auth-httplib2